requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
distro==1.8.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.0.6
langchain==0.0.341
langsmith==0.0.67
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.5
packaging==23.2
pip==23.3.1
pydantic-core==2.14.5
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
setuptools==65.5.1
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==2.1.0
wheel==0.38.4
yarl==1.9.3

fastapi~=0.104.1
streamlit~=1.29.0